requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
tushare>=1.3.0
akshare>=1.12.0  # 同花顺/概念板块数据（免费）

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Database
sqlalchemy>=2.0.0

# Utils
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic-settings>=2.0.0

# Optional (Visualization)
# streamlit>=1.28.0
# plotly>=5.18.0
# matplotlib>=3.8.0

# Optional (Technical Analysis)
# TA-Lib>=0.4.28

# Development
pytest>=7.4.0
jupyter>=1.0.0
